Add informer for Service definitions
- adds informer for Service definitions to fix #61 and #60 -
since it seemed that deleting Service definitions was not tracked
or synchronized and the code was depending on the Deployment
from being deleted or updated.

Tested on K8s 1.10 on single node Linux cluster by deleting
Service and seeing it re-created in the logs. Happy path of
deploying from store also works unaffected.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>