Fix rate limiting latency for list endpoint
- the list endpoint of the REST API was making too many API calls
in a short period of time (1 per existing function in cluster)
so we hit rate limiting and the Kubernetes API server started to
rate limit us causing a huge delay to the end-user and timeouts.

This issue was found when Bart S tried out OpenFaas Cloud with
the Operator which makes use of the REST API. The UI would have
had the same issue causing a poor and confusing experience for
our users.

An initial fix would have used List instead of Get to reduce API
calls, but this fix follows an idea from Stefan to make use of
a shared Informer which already has a cache of the state of
Deployments in the cluster.

Tested on Linux Kubernetes 1.10 cluster and could not reproduce
reported issue.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>